+++ packages/components/popper/src/index.ts
@@ -1,12 +1,32 @@
 export {
   OkuPopper,
-  OkuPopperAnchor,
-  OkuPopperContent,
-  OkuPopperArrow,
 } from './popper'
 
 export type {
   PopperRef,
-  PopperAnchorProps,
   PopperProps,
 } from './popper'
+
+export {
+  OkuPopperAnchor,
+} from './popperAnchor'
+
+export type {
+  PopperAnchorProps,
+} from './popperAnchor'
+
+export {
+  OkuPopperContent,
+} from './popperContent'
+
+export type {
+  PopperContentProps,
+} from './popperContent'
+
+export {
+  OkuPopperArrow,
+} from './popperArrow'
+
+export type {
+  PopperArrowProps,
+} from './popperArrow'
